Separate out components for EditEntries>Group
