+++ src/views/EditEntries/Group/EntryGroupItem/LabelItem/styles.scss
@@ -1,0 +1,38 @@
+.label-item {
+    display: flex;
+    position: relative;
+    flex-basis: calc(100% / 3);
+    flex-direction: column;
+    margin: var(--spacing-small);
+    background-color: var(--color-foreground);
+    padding: var(--spacing-small);
+    min-height: 120px;
+
+    .preview-title {
+        display: flex;
+        flex-direction: row;
+        flex-shrink: 0;
+        padding: var(--spacing-small);
+
+        .heading {
+            flex-grow: 1;
+        }
+
+        .button {
+            opacity: 0;
+        }
+    }
+
+    .entry-preview {
+        flex-grow: 1;
+        padding: var(--spacing-small);
+    }
+
+    &:hover {
+        .preview-title {
+            .button {
+                opacity: 1;
+            }
+        }
+    }
+}
