+++ src/views/EditEntries/Group/EntryGroupItem/styles.scss
@@ -1,0 +1,26 @@
+.entry-group-item {
+    display: flex;
+    flex-direction: column;
+    padding: var(--spacing-small);
+
+    .label-header {
+        display: flex;
+        flex-direction: row;
+        padding: var(--spacing-small);
+
+        .heading {
+            align-self: center;
+            flex-grow: 1;
+        }
+
+        .button {
+            flex-shrink: 0;
+            margin-left: var(--spacing-small);
+        }
+    }
+
+    .label-container {
+        display: flex;
+        flex-direction: row;
+    }
+}
