+++ src/views/EditEntries/Group/styles.scss
@@ -43,72 +43,6 @@
             flex-direction: column;
             flex-grow: 1;
             overflow: auto;
-
-            .entry-group-item {
-                display: flex;
-                flex-direction: column;
-                padding: var(--spacing-small);
-
-                .label-header {
-                    display: flex;
-                    flex-direction: row;
-                    padding: var(--spacing-small);
-
-                    .heading {
-                        align-self: center;
-                        flex-grow: 1;
-                    }
-
-                    .button {
-                        flex-shrink: 0;
-                        margin-left: var(--spacing-small);
-                    }
-                }
-
-                .label-container {
-                    display: flex;
-                    flex-direction: row;
-
-                    .label-item {
-                        display: flex;
-                        position: relative;
-                        flex-basis: calc(100% / 3);
-                        flex-direction: column;
-                        margin: var(--spacing-small);
-                        background-color: var(--color-foreground);
-                        padding: var(--spacing-small);
-                        min-height: 120px;
-
-                        .preview-title {
-                            display: flex;
-                            flex-direction: row;
-                            flex-shrink: 0;
-                            padding: var(--spacing-small);
-
-                            .heading {
-                                flex-grow: 1;
-                            }
-
-                            .button {
-                                opacity: 0;
-                            }
-                        }
-
-                        .entry-preview {
-                            flex-grow: 1;
-                            padding: var(--spacing-small);
-                        }
-
-                        &:hover {
-                            .preview-title {
-                                .button {
-                                    opacity: 1;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
         }
     }
 }
